packages/poly_bench/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "poly-bench"
version = "1.0.0"
description = "Translation model comparison and benchmarking tools"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Enro Team"}
]
keywords = ["translation", "benchmark", "comparison", "bleu", "evaluation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    "pyyaml>=6.0",
    "nltk>=3.8",
    # For monorepo: imports from src.poly_trans
    # For PyPI release: "poly-trans>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "mypy>=1.0",
]

[tool.setuptools]
packages = ["poly_bench"]
